// back-end/api-gateway-example/lib/util/schemas/employee-zod-schema.ts
export type createEmployeeZodSchema = {
  last_name: string;
  first_name: string;
  title: string;
  title_of_courtesy: string;
  birth_date: string;
  hire_date: string;
  address: string;
  city: string;
  region: string;
  postal_code: string;
  country: string;
  home_phone: string;
  extension: string;
  photo: Buffer;
  notes: string;
  reports_to: number;
  photo_path: string;
  territory_id: string;
};